+++ packages/core/src/interface.ts
@@ -26,4 +26,5 @@ export interface Cache {
   get: (location: HickoryLocation) => AnyResponse;
 }
 
-export type Params = { [key: string]: string };
+export type RawParams = { [key: string]: string };
+export type Params = { [key: string]: any };