+++ packages/core/types/index.d.ts
@@ -1,4 +1,4 @@
-export { AddonRegister, AddonGet, Addon, AddonFactory, Subscriber, SideEffect, UnsubscribeFn, Cache, RawParams, Params } from './interface';
+export { AddonRegister, AddonGet, Addon, Subscriber, SideEffect, UnsubscribeFn, Cache, RawParams, Params } from './interface';
 export { Path } from './utils/createPath';
 export { Route, RouteDescriptor, Title, ParamParser, ParamParsers, LoadModifiers, LoadFn, PreloadFn } from './utils/createRoute';
 export { Response, RedirectResponse, AnyResponse } from './utils/createResponse';