+++ packages/core/types/interface.d.ts
@@ -9,7 +9,6 @@ export interface Addon {
     get: AddonGet;
     reset(): void;
 }
-export declare type AddonFactory = (...args: Array<any>) => Addon;
 export declare type Subscriber = (response: AnyResponse, action?: string) => void;
 export interface SideEffect {
     fn: Subscriber;